refactor(services/shop): cambie paths para la correcta importación de middlewares
