+++ services/shop/src/v1/router/inputs/loginLogout.js
@@ -1,5 +1,5 @@
 import { Router } from 'express'
-import { passport, isAuthenticated, login } from '../../controllers/services/authPassport.js'
+import { passport, isAuthenticated, login } from '../../../../../utils/middlewares/auth/authPassport.js'
 
 const logInOut = Router()
