+++ services/shop/src/v1/router/inputs/register.js
@@ -1,5 +1,7 @@
 import { Router } from 'express'
 import register from '../../controllers/user/register.js'
+
+
 const registerRouter = Router()
 
 registerRouter.post('/register', register)
